--- packages/@molt/command/src/extensions/zod/TypeAdaptor/TypeAdaptors.ts
@@ -1,1 +1,0 @@
-export * from './zod/index.js'
