--- packages/@molt/command/src/extensions/zod/TypeAdaptor/index.ts
@@ -1,1 +1,0 @@
-export * as TypeAdaptor from './TypeAdaptors.js'
