--- packages/@molt/command/src/extensions/zod/TypeAdaptor/zod/index.ts
@@ -1,1 +1,0 @@
-export * as Zod from './zod.js'
