+++ packages/@molt/command/src/extensions/zod/guards.ts
