--- packages/@molt/command/src/extensions/zod/zodHelpers/index.ts
@@ -1,1 +1,0 @@
-export * as ZodHelpers from './index_.js'
